--- src/main/java/Command.java
@@ -1,16 +1,0 @@
-abstract class Command {
-
-    private final String command;
-
-    Command(String command) {
-        this.command = command;
-    }
-
-    String getCommand() {
-        return this.command;
-    }
-
-    abstract void execute(TaskRecords taskList, BotUI ui) throws DukeException;
-
-    abstract boolean isExit();
-}
