+++ src/main/java/duke/DukeException.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class DukeException extends Exception {
     public DukeException(String errorMsg) {
         super(errorMsg);
