+++ src/main/java/duke/Echo.java
@@ -1,0 +1,13 @@
+package duke;
+public class Echo {
+    private final String echoText;
+
+    Echo(String echoText) {
+        this.echoText = echoText;
+    }
+
+    @Override
+    public String toString() {
+        return this.echoText;
+    }
+}
