+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,24 +1,27 @@
-public class MarkCommand extends Command {
+package duke.command;
+
+import duke.storage.TaskRecords;
+import duke.ui.BotUI;
+import duke.DukeException;
+public class DeleteCommand extends Command {
     private final String details;
 
-    MarkCommand(String command, String details) {
+    public DeleteCommand(String command, String details) {
         super(command);
         this.details = details;
     }
 
     @Override
-    public void execute(TaskRecords taskList, BotUI ui) throws DukeException {
+    public void execute(TaskRecords taskList, BotUI ui) throws DukeException{
         try {
             int taskIdx = Integer.parseInt(details) - 1;
-            Task currTask = taskList.get(taskIdx);
-            currTask = (super.getCommand().equals("mark")) ? currTask.markDone() : currTask.unmarkDone();
-            taskList.addProcess(currTask);
-            System.out.print(ui.informMarkStatus(taskList.get(taskIdx)));
+            System.out.print(ui.successRemoved(taskList, taskList.delete(taskIdx)));
         } catch (NumberFormatException ex) {
             throw new DukeException(ui.invalidCheckFormat()) ;
         } catch (IndexOutOfBoundsException ex) {
             throw new DukeException(ui.taskNotExist(taskList));
         }
+
     }
 
     @Override
